MLOps-Pipeline/requirements.txt
# MLOps Pipeline CDK Dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0
sagemaker>=2.200.0
boto3>=1.28.0
